+++ snk.c
@@ -33,7 +33,7 @@ snk_position_compare(const snk_position *a, const snk_position *b)
 }
 
 static int
-snk_check_position_possible(const snk_position *position, uint8_t field_width, uint8_t field_height)
+snk_check_position_possible(const snk_position *position, uint32_t field_width, uint32_t field_height)
 {
     if (position->x < 0 || position->x >= field_width)
         return EINVAL;
@@ -57,7 +57,7 @@ static int
 snk_check_position_available(const snk_position *position, const snk_field *field)
 {
     int rc;
-    uint8_t i;
+    uint32_t i;
 
     rc = snk_check_position_possible(position, field->width, field->height);
     if (rc != 0)
@@ -74,7 +74,7 @@ snk_check_position_available(const snk_position *position, const snk_field *fiel
 }
 
 static int
-snk_check_field_obstacle_possible(const snk_field_obstacle *obstacle, uint8_t field_width, uint8_t field_height)
+snk_check_field_obstacle_possible(const snk_field_obstacle *obstacle, uint32_t field_width, uint32_t field_height)
 {
     int rc;
 
@@ -86,7 +86,7 @@ snk_check_field_obstacle_possible(const snk_field_obstacle *obstacle, uint8_t fi
 }
 
 int
-snk_create_field(uint8_t width, uint8_t height, uint8_t n_obstacles, const snk_field_obstacle *obstacles,
+snk_create_field(uint32_t width, uint32_t height, uint32_t n_obstacles, const snk_field_obstacle *obstacles,
                  uint32_t rand_seed, snk_field *field)
 {
     size_t i;
@@ -117,14 +117,14 @@ snk_create_field(uint8_t width, uint8_t height, uint8_t n_obstacles, const snk_f
 struct snk_check_snake_data {
     const snk_field *field;
     snk_position positions[64];
-    uint8_t n_positions;
+    uint32_t n_positions;
 };
 
 static int
 snk_check_snake_cb(const snk_position *pos, void *data)
 {
     struct snk_check_snake_data *check_data = data;
-    uint8_t i;
+    uint32_t i;
     int rc;
 
     rc = snk_check_position_available(pos, check_data->field);
@@ -153,7 +153,7 @@ snk_check_snake(const snk_snake *snake, const snk_field *field)
 
 void
 snk_snake_init(const snk_position *pos, snk_direction direction, const snk_joint_buffer *joints,
-               uint16_t length, uint16_t pending_length, snk_snake *snake)
+               uint32_t length, uint32_t pending_length, snk_snake *snake)
 {
     if (joints != NULL)
         snake->joints = *joints;
@@ -168,7 +168,7 @@ snk_snake_init(const snk_position *pos, snk_direction direction, const snk_joint
 
 int
 snk_create(const snk_field *field, const snk_position *start_position,
-           snk_direction start_direction, uint16_t start_length, snk_process *process)
+           snk_direction start_direction, uint32_t start_length, snk_process *process)
 {
     snk_process result;
     int rc;
@@ -354,7 +354,7 @@ snk_render(const snk_process *process, uint8_t *data, size_t data_size)
     size_t i;
     int rc;
 
-    if ((size_t)(process->field.height * process->field.width) > data_size)
+    if ((process->field.height * process->field.width) > data_size)
         return EPERM;
 
     for (i = 0; i < data_size; i++)
