+++ snk.h
@@ -22,12 +22,12 @@ typedef struct snk_field_obstacle {
 
 /** Field dimensions */
 typedef struct snk_field {
-    uint8_t width;
-    uint8_t height;
+    uint32_t width;
+    uint32_t height;
     snk_field_obstacle obstacles[SNK_FIELD_OBSTACLE_MAX];
-    uint8_t n_obstacles;
+    uint32_t n_obstacles;
     snk_position food;
-    uint8_t n_food;
+    uint32_t n_food;
     uint32_t rand_seed;
 } snk_field;
 
@@ -40,10 +40,10 @@ typedef struct snk_process {
 int snk_position_advance(snk_position *position, snk_direction direction);
 int snk_position_compare(const snk_position *a, const snk_position *b);
 
-int snk_create_field(uint8_t width, uint8_t height, uint8_t n_obstacles, const snk_field_obstacle *obstacles,
+int snk_create_field(uint32_t width, uint32_t height, uint32_t n_obstacles, const snk_field_obstacle *obstacles,
                      uint32_t rand_seed,  snk_field *field);
 int snk_create(const snk_field *field, const snk_position *start_position,
-               snk_direction start_direction, uint16_t start_length, snk_process *process);
+               snk_direction start_direction, uint32_t start_length, snk_process *process);
 int snk_next_tick(snk_process *process);
 int snk_choose_direction(snk_process *process, snk_direction direction);
 int snk_render(const snk_process *process, uint8_t *data, size_t data_size);
